POC to reuse MenuBlock
